normalize ports for non-icmp protocols
for rules with custom protocols, when make_rule_key() is run from
inside addRulesToLookup(), the ports in the key are set to None.  when
make_rule_key() is run later to compare the ec2_group module arguments,
the ports in the key are set to "-1" or whatever the user entered.  this
commit sets the ports to None so both usages of make_rule_key() will
match up with each other.